1.28 More accurate HOT/MOT/DOT => abnormality mapping. Add remote debug output to check whether rootowner is needed for it. Project references and config cleanup.
