+++ Data/packages.config
@@ -1,10 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-
 <packages>
-  <package id="EntityFramework" version="6.0.0" targetFramework="net40" />
-  <package id="log4net" version="2.0.5" targetFramework="net4" />
-  <package id="System.Data.SQLite" version="1.0.101.0" targetFramework="net40" />
-  <package id="System.Data.SQLite.Core" version="1.0.101.0" targetFramework="net40" />
-  <package id="System.Data.SQLite.EF6" version="1.0.101.0" targetFramework="net40" />
-  <package id="System.Data.SQLite.Linq" version="1.0.101.0" targetFramework="net40" />
+  <package id="log4net" version="2.0.5" targetFramework="net46" />
 </packages>
